uefi_app: split bsa acs for DT vs ACPI
BSA for IR or ES band each link to a different BsaPalLib. Instead of
building an application to check either, build one for each case linking
with the appropriate PAL library.

Instead of Bsa.efi, there is now BsaDT.efi and BsaACPI.efi. This way
there is no need to patch the sources and both certification tools can
be part of an ACS image.

Signed-off-by: Eric Chanudet <echanude@redhat.com>